backend/requirements.txt
Django==1.9.4
psycopg2==2.6.1
djangorestframework==3.3.3
djangorestframework-jwt==1.8
django-cors-headers==1.1.0
djangorestframework-camel-case==0.2.0
django-model-utils==2.5
requests==2.9.1
